% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.r, R/sankey.r
\docType{methods}
\name{chordMovePlot}
\alias{chordMovePlot}
\alias{sankeyMovePlot}
\title{chordMovePlot}
\usage{
chordMovePlot(obj, ...)

sankeyMovePlot(obj, ...)
}
\arguments{
\item{age:}{By default the first age is plotted}

\item{season:}{By default the first season is plotted}

\item{obj:}{An object either of class MFCLRegion or of class MFCLTag}

\item{frq:}{An object of MFCLFrq. Defaults to NULL but required if obj=MFCLTag}
}
\value{
A plot of movement between assessment regions as determined either by the tag data or the estimated movement from the assessment.

A plot of movement between assessment regions as determined either by the tag data or the estimated movement from the assessment.
}
\description{
Produces a Chord plot of movement

Produces a Sankey plot of movement
}
